chore: use appforge parent project, code refactoring
