+++ src/main/java/co/bitshifted/appforge/ignite/deploy/DependencyResolutionResult.java
@@ -1,6 +1,6 @@
 /*
  *
- *  * Copyright (c) 2022  Bitshift D.O.O (http://bitshifted.co)
+ *  * Copyright (c) 2022-2022  Bitshift D.O.O (http://bitshifted.co)
  *  *
  *  * This Source Code Form is subject to the terms of the Mozilla Public
  *  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -8,10 +8,10 @@
  *
  */
 
-package co.bitshifted.ignite.deploy;
+package co.bitshifted.appforge.ignite.deploy;
 
-import co.bitshifted.ignite.common.model.OperatingSystem;
-import co.bitshifted.ignite.model.JavaDependency;
+import co.bitshifted.appforge.common.model.OperatingSystem;
+import co.bitshifted.appforge.ignite.model.JavaDependency;
 
 import java.util.HashSet;
 import java.util.Set;
