+++ src/main/java/co/bitshifted/appforge/ignite/exception/CommunicationException.java
@@ -1,6 +1,6 @@
 /*
  *
- *  * Copyright (c) 2022  Bitshift D.O.O (http://bitshifted.co)
+ *  * Copyright (c) 2022-2022  Bitshift D.O.O (http://bitshifted.co)
  *  *
  *  * This Source Code Form is subject to the terms of the Mozilla Public
  *  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -8,7 +8,7 @@
  *
  */
 
-package co.bitshifted.ignite.exception;
+package co.bitshifted.appforge.ignite.exception;
 
 public class CommunicationException extends Exception{
