+++ src/main/java/co/bitshifted/appforge/ignite/http/SubmitDeploymentResponse.java
@@ -1,6 +1,6 @@
 /*
  *
- *  * Copyright (c) 2022  Bitshift D.O.O (http://bitshifted.co)
+ *  * Copyright (c) 2022-2022  Bitshift D.O.O (http://bitshifted.co)
  *  *
  *  * This Source Code Form is subject to the terms of the Mozilla Public
  *  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -8,9 +8,9 @@
  *
  */
 
-package co.bitshifted.ignite.http;
+package co.bitshifted.appforge.ignite.http;
 
-import co.bitshifted.ignite.common.dto.RequiredResourcesDTO;
+import co.bitshifted.appforge.common.dto.RequiredResourcesDTO;
 import lombok.Data;
 
 @Data
