+++ src/main/java/co/bitshifted/appforge/ignite/model/JavaDependency.java
@@ -1,6 +1,6 @@
 /*
  *
- *  * Copyright (c) 2022  Bitshift D.O.O (http://bitshifted.co)
+ *  * Copyright (c) 2022-2022  Bitshift D.O.O (http://bitshifted.co)
  *  *
  *  * This Source Code Form is subject to the terms of the Mozilla Public
  *  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -8,10 +8,10 @@
  *
  */
 
-package co.bitshifted.ignite.model;
+package co.bitshifted.appforge.ignite.model;
 
-import co.bitshifted.ignite.common.dto.JavaDependencyDTO;
-import co.bitshifted.ignite.util.ModuleChecker;
+import co.bitshifted.appforge.common.dto.JavaDependencyDTO;
+import co.bitshifted.appforge.ignite.util.ModuleChecker;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.*;
