+++ src/main/java/co/bitshifted/appforge/ignite/resource/ResourceProducer.java
@@ -1,6 +1,6 @@
 /*
  *
- *  * Copyright (c) 2022  Bitshift D.O.O (http://bitshifted.co)
+ *  * Copyright (c) 2022-2022  Bitshift D.O.O (http://bitshifted.co)
  *  *
  *  * This Source Code Form is subject to the terms of the Mozilla Public
  *  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -8,10 +8,10 @@
  *
  */
 
-package co.bitshifted.ignite.resource;
+package co.bitshifted.appforge.ignite.resource;
 
 
-import co.bitshifted.ignite.common.model.BasicResource;
+import co.bitshifted.appforge.common.model.BasicResource;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -20,7 +20,7 @@ import java.nio.file.Paths;
 import java.util.Collections;
 import java.util.List;
 
-import static co.bitshifted.ignite.IgniteConstants.*;
+import static co.bitshifted.appforge.ignite.IgniteConstants.*;
 
 public class ResourceProducer {
