+++ src/main/java/co/bitshifted/appforge/ignite/util/ModuleChecker.java
@@ -1,6 +1,6 @@
 /*
  *
- *  * Copyright (c) 2022  Bitshift D.O.O (http://bitshifted.co)
+ *  * Copyright (c) 2022-2022  Bitshift D.O.O (http://bitshifted.co)
  *  *
  *  * This Source Code Form is subject to the terms of the Mozilla Public
  *  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -8,7 +8,7 @@
  *
  */
 
-package co.bitshifted.ignite.util;
+package co.bitshifted.appforge.ignite.util;
 
 
 import org.apache.maven.plugin.logging.Log;
