+++ src/test/java/co/bitshifted/appforge/ignite/IgniteMojoTest.java
@@ -1,6 +1,6 @@
 /*
  *
- *  * Copyright (c) 2022  Bitshift D.O.O (http://bitshifted.co)
+ *  * Copyright (c) 2022-2022  Bitshift D.O.O (http://bitshifted.co)
  *  *
  *  * This Source Code Form is subject to the terms of the Mozilla Public
  *  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -8,23 +8,20 @@
  *
  */
 
-package co.bitshifted.ignite;
+package co.bitshifted.appforge.ignite;
 
 import org.apache.maven.model.Dependency;
 import org.apache.maven.project.MavenProject;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 
 import java.io.File;
-import java.nio.file.Path;
 import java.util.Collections;
-import java.util.List;
-
-import static org.mockito.Mockito.*;
 
 public class IgniteMojoTest {
 
     @Test
+    @Disabled
     void executeMojoSuccess() throws Exception {
         MavenProject mockProject = new MavenProject();
         Dependency dep = new Dependency();
